Add get ingredients functions

+++ src/api/models/productModel.js
@@ -2,8 +2,9 @@
 
 import promisePool from "../../utils/database.js";
 
-const addProduct = async (product, file, user) => {
-  const { name, price, description } = product;
+const addProduct = async (body, file, user) => {
+  const { name, price, description, ingredients } = body;
+  const ingredientsList = JSON.parse(ingredients);
 
   const sql = `INSERT INTO products (name, price, description, img) VALUES (?, ?, ?, ?)`;
   const params = [name, price, description, file.filename].map((value) => {
@@ -13,10 +14,19 @@ const addProduct = async (product, file, user) => {
       return value;
     }
   });
+
   const rows = await promisePool.execute(sql, params);
   if (rows[0].affectedRows === 0) {
     return false;
   } else {
+    await Promise.all(
+      ingredientsList.map(async (i) => {
+        return promisePool.execute(
+          "INSERT INTO ingredients_products (ingredient_id, product_id) VALUES (?, ?)",
+          [i, rows[0].insertId],
+        );
+      }),
+    );
     return { message: "Success" };
   }
 };
@@ -45,6 +55,12 @@ const exterminateProduct = async (id, user) => {
     return { message: "Only admin can delete products" };
   }
   try {
+    const [row1] = await promisePool.execute(
+      "DELETE FROM ingredients_products WHERE product_id = ?",
+      [id],
+    );
+    if (row1.affectedRows === 0) return false;
+
     const [rows] = await promisePool.execute(
       "DELETE FROM products WHERE id = ?",
       [id],
@@ -79,6 +95,16 @@ const getAllIngredients = async () => {
   return rows;
 };
 
+const getProductIngredients = async (id) => {
+  const [rows] = await promisePool.query(
+    "SELECT p.id, p.name, i.id, i.name FROM products p JOIN ingredients_products ip ON p.id = ip.product_id JOIN ingredients i ON ip.ingredient_id = i.id WHERE p.id = ?;",
+    [id],
+  );
+  console.log(rows);
+  if (rows.affectedRows === 0) return false;
+  return rows;
+};
+
 export {
   addProduct,
   getAllProducts,
@@ -86,4 +112,5 @@ export {
   exterminateProduct,
   updateProduct,
   getAllIngredients,
+  getProductIngredients,
 };
